brainscore_vision/models/temporal_model_openstl/requirements.txt
torchvision
openstl @ git+https://github.com/YingtianDt/OpenSTL.git